Start and stop recording using a macro
The previous code didn't work as PyATEMMax doesn't have a function to handle Atem ISO Recordings.
I found a way around it creating two custom macros to start and stop recording.
That code works but I will try to find a more elegant way to do it in the future.